+++ src/libs/common/index.ts
@@ -32,4 +32,6 @@ export * from "./types/TextEditorDecorationType";
 export * from "./types/TextEditorEdit";
 export * from "./types/TextEditorOptions";
 export * from "./types/TextLine";
+export * from "./testUtil/fromPlainObject";
+export * from "./testUtil/toPlainObject";
 export { default as DefaultMap } from "./util/DefaultMap";
