+++ src/libs/vscode-common/index.ts
@@ -1,6 +1,6 @@
 export * from "./getExtensionApi";
 export * from "./notebook";
-export * from "./testUtil/toPlainObject";
+export * from "../common/testUtil/toPlainObject";
 export * from "./testUtil/testConstants";
 export * from "./testUtil/openNewEditor";
 export * from "./vscodeUtil";
